+++ src/app/categories/components/AddInitCategory/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AddEditCategory';
