+++ src/app/categories/components/AddInitCategory/types.ts
@@ -1,0 +1,5 @@
+import type { CategoryType } from '@app/types/list.types';
+
+export interface AddEditProps {
+  category: CategoryType | null;
+}
