--- src/app/categories/components/categoryCard/classes.module.css
@@ -1,28 +1,0 @@
-.main-container {
-  display: flex;
-  width: 60%;
-  margin-top: 2rem;
-}
-.buttons {
-  display: flex;
-  width: 40%;
-  justify-content: space-around;
-}
-
-@media (max-width: 768px) {
-  .main-container {
-    flex-direction: column;
-    margin-top: 40px;
-    width: 90%;
-  }
-  .title {
-    align-self: self-start;
-  }
-  .buttons {
-    width: 100%;
-    align-items: self-start;
-    justify-content: start;
-    gap: 20px 30px;
-    margin-top: 20px;
-  }
-}
