+++ src/app/categories/components/categoryCard/types.ts
@@ -2,9 +2,7 @@ import type { CategoryType } from '@app/types/list.types';
 export interface CategoryCardProps {
   category: CategoryType;
 }
-export interface AddEditProps {
-  category: CategoryType | null;
-}
+
 export interface CategoryParamProp {
   categoryId: string;
 }
