+++ src/app/categories/styles.module.css
@@ -1,0 +1,28 @@
+.main {
+  width: 100%;
+  max-height: calc(100vh - var(--header-height) - var(--footer-height));
+  padding-bottom: 20px;
+  overflow: scroll;
+}
+
+.wrapper {
+  flex-direction: column;
+  align-items: center;
+  max-width: var(--max-layout-width);
+  margin: 0 auto;
+  width: 100%;
+}
+
+@media (min-width: 768px) {
+  .main {
+    max-height: calc(
+      100vh - var(--header-height-tablet) - var(--footer-height)
+    );
+  }
+}
+
+@media (min-width: 1024px) {
+  .main {
+    padding-bottom: 20px;
+  }
+}
