+++ src/app/globals.css
@@ -5,9 +5,12 @@
 
   --footer-height: 50px;
   --header-height: 50px;
+  --header-height-tablet: 80px;
 
   --breakpoint-tablet: 768px;
   --breakpoint-desktop: 1024px;
+
+  --max-layout-width: 1200px;
 }
 
 * {
@@ -20,7 +23,7 @@
 html,
 body {
   max-width: 100vw;
-  overflow-x: hidden;
+  overflow: hidden;
 }
 
 body {
