+++ src/app/list/[id]/styles.module.css
@@ -1,0 +1,13 @@
+.main {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+  padding-top: 10px;
+}
+
+@media (min-width: 768px) {
+  .main {
+    padding-top: 40px;
+  }
+}
