+++ src/app/list/components/ListTitle/styles.module.css
@@ -8,3 +8,15 @@
   font-size: 19px;
   text-align: center;
 }
+
+@media (min-width: 768px) {
+  .list-title-input {
+    font-size: 24px;
+  }
+}
+
+@media (min-width: 1024px) {
+  .list-title-input {
+    font-size: 28px;
+  }
+}
