+++ src/app/list/components/itemsForm/components/createItem/styles.module.css
@@ -14,3 +14,25 @@
   width: 100%;
   padding: 2px 10px;
 }
+
+@media (min-width: 768px) {
+  .create-input-wrapper {
+    padding-bottom: 24px;
+    min-width: 400px;
+  }
+
+  .create-input {
+    font-size: 24px;
+  }
+}
+
+@media (min-width: 1024px) {
+  .create-input-wrapper {
+    padding-bottom: 32px;
+    min-width: 600px;
+  }
+
+  .create-input {
+    font-size: 28px;
+  }
+}
