+++ src/app/main/mainPage.module.css
@@ -9,14 +9,56 @@
 
 .filter {
   font-family: var(--font-allura);
+  max-width: var(--max-layout-width);
+  margin: 0 auto;
 }
 
 .main-wrapper {
   padding: 20px;
+  max-height: calc(100vh - var(--footer-height) - var(--header-height));
+  overflow: scroll;
+  width: 100%;
+}
+
+.grid {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  grid-column-gap: 12px;
+  grid-row-gap: 12px;
+  width: auto;
 }
 
 @media (min-width: 768px) {
   .main-wrapper {
-    padding: 40px;
+    padding-top: 40px;
+    max-height: calc(
+      100vh - var(--footer-height) - var(--header-height-tablet)
+    );
+  }
+
+  .grid {
+    grid-column-gap: 20px;
+    grid-row-gap: 20px;
+  }
+
+  .filter {
+    font-size: 24px;
+  }
+}
+
+@media (min-width: 1024px) {
+  .grid {
+    max-width: var(--max-layout-width);
+    margin: 0 auto;
+    grid-template-columns: repeat(4, 1fr);
+    grid-column-gap: 30px;
+    grid-row-gap: 30px;
+    padding: 0 40px;
+  }
+
+  .filter {
+    padding: 0 40px;
+    margin-bottom: 20px;
+    font-size: 32px;
   }
 }
