+++ src/components/editAndSaveButton/styles.module.css
@@ -3,3 +3,22 @@
   height: 24px;
   padding: 2px;
 }
+
+@media (min-width: 768px) {
+  .button {
+    width: 32px;
+    height: 32px;
+  }
+
+  .icon {
+    width: 24px;
+    height: 24px;
+  }
+}
+
+@media (min-width: 1024px) {
+  .button {
+    width: 40px;
+    height: 40px;
+  }
+}
