+++ src/components/header/header.module.css
@@ -28,13 +28,15 @@
   }
 
   .text {
-    font-size: 32px;
+    font-size: 42px;
   }
 }
 
 @media (min-width: 1024px) {
   .header {
     justify-content: space-between;
+    max-width: var(--max-layout-width);
+    margin: 0 auto;
   }
 
   .text {
